+++ src/lib.rs
@@ -18,7 +18,6 @@ mod git_repo_enumerator;
 pub mod git_url;
 pub mod github;
 pub mod gitlab;
-pub mod guesser;
 pub mod jira;
 pub mod liquid_filters;
 pub mod location;